+++ src/Exception/InvalidArgument.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LosMiddleware\RateLimit\Exception;
+
+class InvalidArgument extends ReachedRateLimit
+{
+}