+++ src/Exception/InvalidArgumentException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace LosMiddleware\RateLimit\Exception;
-
-class InvalidArgumentException extends RateLimitException
-{
-}