+++ src/Exception/MissingParameterException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace LosMiddleware\RateLimit\Exception;
-
-class MissingParameterException extends RateLimitException
-{
-}