+++ src/Storage/StorageInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace LosMiddleware\RateLimit\Storage;
-
-interface StorageInterface
-{
-    public function get($key, $default = 0);
-    public function set($key, $value);
-}