+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## 3.0.0 - 2019-07-15
+
+### Added
+
+- Support for psr/simple-cache as storage
+
+### Changed
+
+- Updated dependencies, specially support for zend-diactoros 1.8 and 2.0
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Internal cache implementation
+
+### Fixed
+
+- Nothing.
+
 ## 2.1.0 - 2018-05-08
 
 ### Added